Tests and bug fixes (#1)
Added integration tests and fixed some bugs

+++ client/src/main.rs
@@ -3,25 +3,41 @@ mod transfer;
 extern crate base64;
 extern crate openssl;
 
-use crate::transfer::transfer::{Download, Upload, TransferError};
+use crate::transfer::transfer::{Download, TransferError, Upload};
 use std::convert::TryInto;
 use std::env::args;
+use std::fs;
 use std::fs::File;
-use std::io::{Read, Write, Seek, SeekFrom, stdin, stdout};
+use std::io::{stdin, stdout, Read, Seek, SeekFrom, Write};
 use std::net::ToSocketAddrs;
-use std::path::{PathBuf, Path};
+use std::path::{Path, PathBuf};
 use std::process::exit;
 use std::time::Instant;
-use std::fs;
 
 const BUFFER_SIZE: usize = 1024 * 1024;
 
+macro_rules! printinfoln {
+    ($q:expr, $($x:expr), *) => {
+        if !$q {
+        println!($($x,)*);
+        }
+    };
+}
+
+macro_rules! printinfo {
+    ($q:expr, $($x:expr), *) => {
+        if !$q {
+        print!($($x,)*);
+        }
+    };
+}
 fn main() {
     let mut args = args();
     args.next();
 
     let mut encrypt = true;
     let mut receive = false;
+    let mut quiet = false;
     let mut main_arg = None;
 
     loop {
@@ -34,6 +50,14 @@ fn main() {
                     encrypt = false;
                 } else if arg == "-r" {
                     receive = true;
+                } else if arg == "--help" {
+                    println!("Usage: sfshr [file] or sfshr -r [download key]");
+                    println!(" -r [download key] - download file");
+                    println!(" -n --no-encryption - do not encrypt or decrypt the file");
+                    println!(" -q --quiet - do not print anything (expect download key)");
+                    exit(0);
+                } else if arg == "-q" || arg == "--quiet" {
+                    quiet = true;
                 } else {
                     main_arg = Some(arg);
                 }
@@ -43,23 +67,23 @@ fn main() {
 
     if !receive {
         if main_arg.is_none() {
-            println!("No file specified!");
+            printinfoln!(quiet, "No file specified!");
             exit(1);
         }
 
         let path = PathBuf::from(main_arg.unwrap());
-        upload("localhost:40788", path, encrypt);
+        upload("localhost:40788", path, encrypt, quiet);
     } else {
         if main_arg.is_none() {
-            println!("No download key specified!");
+            printinfoln!(quiet, "No download key specified!");
             exit(1);
         }
         let download_key = base64::decode(main_arg.unwrap().as_bytes());
         if download_key.is_err() {
-            println!("Invalid download key format!");
+            printinfoln!(quiet, "Invalid download key format!");
             exit(1);
         }
-        download("localhost:40788", download_key.unwrap(), encrypt);
+        download("localhost:40788", download_key.unwrap(), encrypt, quiet);
     }
 }
 
@@ -92,7 +116,7 @@ impl FormatSize for f64 {
 
         let mut i: f64 = 1.0;
         let mut index = 0;
-        if !self.is_normal(){
+        if !self.is_normal() {
             return String::from("NaN");
         }
         loop {
@@ -108,10 +132,10 @@ impl FormatSize for f64 {
     }
 }
 
-fn upload<A: ToSocketAddrs>(addr: A, filepath: PathBuf, encrypt: bool) {
+fn upload<A: ToSocketAddrs>(addr: A, filepath: PathBuf, encrypt: bool, quiet: bool) {
     let file = File::open(&filepath);
     if file.is_err() {
-        println!("Failed to open file: {}", file.err().unwrap());
+        printinfoln!(quiet, "Failed to open file: {}", file.err().unwrap());
         exit(1);
     }
 
@@ -119,7 +143,9 @@ fn upload<A: ToSocketAddrs>(addr: A, filepath: PathBuf, encrypt: bool) {
 
     let mut upload = Upload::new(addr, encrypt).unwrap_or_else(on_error);
 
-    upload.write_filename(filepath.file_name().unwrap().to_str().unwrap()).unwrap_or_else(on_error);
+    upload
+        .write_filename(filepath.file_name().unwrap().to_str().unwrap())
+        .unwrap_or_else(on_error);
     let mut buf = Vec::new();
     buf.resize(BUFFER_SIZE, 0);
 
@@ -133,19 +159,26 @@ fn upload<A: ToSocketAddrs>(addr: A, filepath: PathBuf, encrypt: bool) {
         upload.send(&buf[..bytes_read]).unwrap_or_else(on_error);
         let time = time.elapsed().as_micros() as f64;
         let size = file.seek(SeekFrom::Current(0)).unwrap();
-        let speed = size as f64/time*1000000.0;
-
-        println!("Uploaded {:^12} @ {:^12}  \x1b[1A\x1b[0G", size.format_size(), format!("{}/s", speed.format_size()));
+        let speed = size as f64 / time * 1000000.0;
+        printinfoln!(
+            quiet,
+            "Uploaded {:^12} @ {:^12}  \x1b[1A\x1b[0G",
+            size.format_size(),
+            format!("{}/s", speed.format_size())
+        );
     }
 
     upload.finalize().unwrap_or_else(on_error);
     let time = time.elapsed().as_micros() as f64;
     let size = file.seek(SeekFrom::Current(0)).unwrap();
-    let speed = size as f64/time*1000000.0;
+    let speed = size as f64 / time * 1000000.0;
 
-    println!("Uploaded {:^12} @ {:^12}  \x1b[1A\x1b[0G", size.format_size(), format!("{}/s",  speed.format_size()));
-
-
+    printinfoln!(
+        quiet,
+        "Uploaded {:^12} @ {:^12}  \x1b[1A\x1b[0G",
+        size.format_size(),
+        format!("{}/s", speed.format_size())
+    );
 
     let mut download_key = Vec::new();
     download_key.extend_from_slice(&upload.id());
@@ -160,9 +193,9 @@ fn upload<A: ToSocketAddrs>(addr: A, filepath: PathBuf, encrypt: bool) {
     }
 }
 
-fn download<A: ToSocketAddrs>(addr: A, download_key: Vec<u8>, encrypt: bool) {
+fn download<A: ToSocketAddrs>(addr: A, download_key: Vec<u8>, encrypt: bool, quiet: bool) {
     if (encrypt && download_key.len() != 64) || (!encrypt && download_key.len() != 32) {
-        println!("Invalid download key size!");
+        printinfoln!(quiet, "Invalid download key size!");
         exit(1);
     }
 
@@ -170,7 +203,8 @@ fn download<A: ToSocketAddrs>(addr: A, download_key: Vec<u8>, encrypt: bool) {
     if encrypt {
         key = Some(download_key[32..].try_into().unwrap());
     }
-    let mut download = Download::new(addr, &download_key[..32].try_into().unwrap(), key).unwrap_or_else(on_error);
+    let mut download =
+        Download::new(addr, &download_key[..32].try_into().unwrap(), key).unwrap_or_else(on_error);
     let mut file = File::create(".sfshr-temp").unwrap();
     let time = Instant::now();
     loop {
@@ -182,38 +216,55 @@ fn download<A: ToSocketAddrs>(addr: A, download_key: Vec<u8>, encrypt: bool) {
         file.write_all(&buf).unwrap();
         let time = time.elapsed().as_micros() as f64;
         let size = file.seek(SeekFrom::Current(0)).unwrap();
-        let speed = size as f64/time*1000000.0;
+        let speed = size as f64 / time * 1000000.0;
 
-        println!("Downloaded {:^12} @ {:^12}\x1b[1A\x1b[0G", size.format_size(), format!("{}/s", speed.format_size()));
+        printinfoln!(
+            quiet,
+            "Downloaded {:^12} @ {:^12}\x1b[1A\x1b[0G",
+            size.format_size(),
+            format!("{}/s", speed.format_size())
+        );
     }
-    println!();
+    printinfoln!(quiet, "");
 
     let filename = download.filename();
     if Path::new(&filename).exists() {
-        loop {
-            print!("\x1b[1A\x1b[0G\x1b[KFile {} already exists. Replace? [yes/no]", filename);
-            stdout().flush().unwrap();
-            let stdin = stdin();
-            let mut line = String::new();
-            stdin.read_line(&mut line).unwrap();
-            line.retain(|c| {
-                c != '\n' && c != '\r'
-            });
-            if line == "yes" {
-                break;
-            }
-            if line == "no" {
-                fs::remove_file(".sfshr-temp").unwrap();
-                println!("\x1b[1A\x1b[0G\x1b[KAborted");
-                exit(1);
+        if !quiet {
+            loop {
+                printinfo!(
+                    quiet,
+                    "\x1b[1A\x1b[0G\x1b[KFile {} already exists. Replace? [yes/no]",
+                    filename
+                );
+                stdout().flush().unwrap();
+                let stdin = stdin();
+                let mut line = String::new();
+                stdin.read_line(&mut line).unwrap();
+                line.retain(|c| c != '\n' && c != '\r');
+                if line == "yes" {
+                    break;
+                }
+                if line == "no" {
+                    fs::remove_file(".sfshr-temp").unwrap();
+                    printinfoln!(quiet, "\x1b[1A\x1b[0G\x1b[KAborted");
+                    exit(1);
+                }
             }
         }
     }
     fs::rename(".sfshr-temp", &filename).unwrap();
-    println!("\x1b[1A\x1b[0G\x1b[KSuccesfully downloaded {}", &filename);
+    printinfoln!(
+        quiet,
+        "\x1b[1A\x1b[0G\x1b[KSuccesfully downloaded {}",
+        &filename
+    );
 }
 
-fn on_error<T>(_: TransferError) -> T {
-    println!("\x1b[31mTerminating due to an error\x1b[0m");
+fn on_error<T>(err: TransferError) -> T {
+    let temp = Path::new(".sfshr-temp");
+    if temp.exists() {
+        fs::remove_file(temp).unwrap();
+    }
+    println!("\x1b[31mTerminating due to an error ({})\x1b[0m", err);
     exit(1);
 }
